fix: resolve clang-tidy, cppcheck, and code warnings
- Add -Wno-unknown-warning-option to clang-tidy (GCC-only flags in
  compile_commands.json: -Wduplicated-cond, -Wlogical-op)
- Fix implicit int→uint16_t conversion in NativeFlowExtractor.cpp:553
- Fix unnecessary QString copy in FilterPanel.cpp:177
- Fix Python DeprecationWarning in cppcheck XML parser (use 'is not None')

+++ src/infra/flow/NativeFlowExtractor.cpp
@@ -550,7 +550,8 @@ bool NativeFlowExtractor::parseEthernetAndIp(
       len - static_cast<std::uint32_t>(kEthernetHeaderSize);
 
   if (etherType == kEtherTypeVlan && payloadLen >= 4) {
-    etherType = (static_cast<std::uint16_t>(payload[2]) << 8) | payload[3];
+    etherType = static_cast<std::uint16_t>(
+        (static_cast<std::uint16_t>(payload[2]) << 8) | payload[3]);
     payload += 4;
     payloadLen -= 4;
   }
